be_pdftool/requirements.txt
fastapi
uvicorn
pikepdf
python-multipart
PyPDF2
pdf2image
pillow
python-docx
python-pptx
pytesseract
pdfminer.six
pdfplumber
pdf2docx
transformers
fastapi
uvicorn
PyPDF2
pytesseract
pdf2image
python-multipart
sqlalchemy
pillow
